+++ src/main/kotlin/icu/windea/pls/lang/config/impl/StellarisSwitchCaseOverriddenConfigProvider.kt
@@ -1,0 +1,5 @@
+package icu.windea.pls.lang.config.impl
+
+class StellarisSwitchCaseOverriddenConfigProvider {
+    //TODO 0.10.3
+}