+++ src/main/kotlin/icu/windea/pls/lang/scope/impl/StellarisComplexTriggerModifierOverriddenScopeContextProvider.kt
@@ -11,7 +11,7 @@ import icu.windea.pls.lang.scope.*
 import icu.windea.pls.script.psi.*
 
 @WithGameType(ParadoxGameType.Stellaris)
-class ParadoxComplexTriggerModifierOverriddenScopeContextProvider : ParadoxOverriddenScopeContextProvider {
+class StellarisComplexTriggerModifierOverriddenScopeContextProvider : ParadoxOverriddenScopeContextProvider {
     override fun getOverriddenScopeContext(element: ParadoxScriptMemberElement, config: CwtDataConfig<*>): ParadoxScopeContext? {
         //重载complex_trigger_modifier = {...}中属性trigger的值对应的作用域上下文
         //兼容trigger_scope的值对应的作用域与当前作用域上下文不匹配的情况